% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_detail_village_202106.R
\docType{data}
\name{age_detail_village_202106}
\alias{age_detail_village_202106}
\title{Data of 2021/06 age and youth statistics by village in Taiwan}
\format{
A tibble with 7,761 rows and 76 variables:
\describe{
  \item{district_code}{dbl id of each district}
  \item{info_time}{chr year and month of data}
  \item{countytown}{chr county(hsien/shih 縣/市 in Chinese) and town(hsiang/chen/shih 鄉/鎮/市 in Chinese) name}
  \item{village}{chr county(tsun/li 村/里 in Chinese) name}
}
}
\source{
\url{https://segis.moi.gov.tw/STAT/Web/Platform/QueryInterface/STAT_QueryProductView.aspx?pid=5E21BD9F68B675D8C72481BBD36AA506&spid=7ED8D58E129BC680}

\url{https://segis.moi.gov.tw/STAT/Web/Platform/QueryInterface/STAT_QueryProductView.aspx?pid=5E21BD9F68B675D8CF0D176F294A86E7&spid=7ED8D58E129BC680}
}
\usage{
age_detail_village_202106
}
\description{
These are combined household registration records collected from SEGIS. Please refer to the link below to check details of columns definition. Columns contain number of people cut in five-year-old age, and number of children/youth.
結合 SEGIS 的兩份資料，共包含:
(1) 五歲年齡組(0-4歲、5-9歲、10-14歲、15-19歲、20-24歲、25-29歲、30-34歲、35-39歲、40-44歲、45-49歲、50-54歲、55-59歲、60-64歲、65-69歲、70-74歲、75-79歲、80-84歲、85-89歲、90-94歲、95-99歲、100歲以上)與性別(小計、男性、女性)人口數交叉統計
(2) 兒童及少年年齡(0-5歲、6-11歲、12-17歲)與性別(小計、男性、女性)人口數交叉統計
}
\keyword{datasets}
