% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\docType{data}
\name{edu_town}
\alias{edu_town}
\title{edu_town}
\format{
data.frame
tibble [368 × 6] (S3: rowwise_df/tbl_df/tbl/data.frame)
$ countytown           : chr [1:368] "高雄市阿蓮區" "高雄市大寮區" "高雄市大社區" "高雄市大樹區" ...
$ population_15up      : num [1:368] 25235 100888 31081 37801 313656 ...
$ university           : num [1:368] 6985 27653 9435 10162 112741 ...
$ university.not       : num [1:368] 18250 73235 21646 27639 200915 ...
$ university.female    : num [1:368] 3363 13039 4381 4900 54478 ...
$ university.not.female: num [1:368] 9120 37201 11195 13280 113946 ...
- attr(*, "groups")= tibble [368 × 1] (S3: tbl_df/tbl/data.frame)
 ..$ .rows: list<int> [1:368] 
 .. ..$ : int 1
 .. ..$ : int 2
 .. ..$ : int 3
 .. ..$ : int 4
 .. ..$ : int 5
 .. ..$ : int 6
 .. ..$ : int 7
 .. ..$ : int 8
 .. ..$ : int 9
 .. ..$ : int 10
 .. ..$ : int 11
 .. ..$ : int 12
 .. ..$ : int 13
 .. ..$ : int 14
 .. ..$ : int 15
 .. ..$ : int 16
 .. ..$ : int 17
 .. ..$ : int 18
 .. ..$ : int 19
 .. ..$ : int 20
 .. ..$ : int 21
 .. ..$ : int 22
 .. ..$ : int 23
 .. ..$ : int 24
 .. ..$ : int 25
 .. ..$ : int 26
 .. ..$ : int 27
 .. ..$ : int 28
 .. ..$ : int 29
 .. ..$ : int 30
 .. ..$ : int 31
 .. ..$ : int 32
 .. ..$ : int 33
 .. ..$ : int 34
 .. ..$ : int 35
 .. ..$ : int 36
 .. ..$ : int 37
 .. ..$ : int 38
 .. ..$ : int 39
 .. ..$ : int 40
 .. ..$ : int 41
 .. ..$ : int 42
 .. ..$ : int 43
 .. ..$ : int 44
 .. ..$ : int 45
 .. ..$ : int 46
 .. ..$ : int 47
 .. ..$ : int 48
 .. ..$ : int 49
 .. ..$ : int 50
 .. ..$ : int 51
 .. ..$ : int 52
 .. ..$ : int 53
 .. ..$ : int 54
 .. ..$ : int 55
 .. ..$ : int 56
 .. ..$ : int 57
 .. ..$ : int 58
 .. ..$ : int 59
 .. ..$ : int 60
 .. ..$ : int 61
 .. ..$ : int 62
 .. ..$ : int 63
 .. ..$ : int 64
 .. ..$ : int 65
 .. ..$ : int 66
 .. ..$ : int 67
 .. ..$ : int 68
 .. ..$ : int 69
 .. ..$ : int 70
 .. ..$ : int 71
 .. ..$ : int 72
 .. ..$ : int 73
 .. ..$ : int 74
 .. ..$ : int 75
 .. ..$ : int 76
 .. ..$ : int 77
 .. ..$ : int 78
 .. ..$ : int 79
 .. ..$ : int 80
 .. ..$ : int 81
 .. ..$ : int 82
 .. ..$ : int 83
 .. ..$ : int 84
 .. ..$ : int 85
 .. ..$ : int 86
 .. ..$ : int 87
 .. ..$ : int 88
 .. ..$ : int 89
 .. ..$ : int 90
 .. ..$ : int 91
 .. ..$ : int 92
 .. ..$ : int 93
 .. ..$ : int 94
 .. ..$ : int 95
 .. ..$ : int 96
 .. ..$ : int 97
 .. ..$ : int 98
 .. ..$ : int 99
 .. .. [list output truncated]
 .. ..@ ptype: int(0)
}
\usage{
edu_town
}
\description{
edu_town
}
\keyword{datasets}
