% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\docType{data}
\name{edu_village_original}
\alias{edu_village_original}
\title{edu_village_original}
\format{
data.frame
tibble [7,760 × 51] (S3: spec_tbl_df/tbl_df/tbl/data.frame)
$ statistic_yyy                              : chr [1:7760] "109" "109" "109" "109" ...
$ district_code                              : chr [1:7760] "65000010001" "65000010002" "65000010003" "65000010004" ...
$ countytown                                 : chr [1:7760] "新北市板橋區" "新北市板橋區" "新北市板橋區" "新北市板橋區" ...
$ village                                    : chr [1:7760] "留侯里" "流芳里" "赤松里" "黃石里" ...
$ edu_age_15up_total                         : chr [1:7760] "1455" "1290" "841" "1049" ...
$ edu_doctor_graduated_m                     : chr [1:7760] "3" "8" "5" "4" ...
$ edu_doctor_graduated_f                     : chr [1:7760] "1" "2" "3" "3" ...
$ edu_doctor_ungraduated_m                   : chr [1:7760] "2" "2" "0" "3" ...
$ edu_doctor_ungraduated_f                   : chr [1:7760] "2" "2" "0" "0" ...
$ edu_master_graduated_m                     : chr [1:7760] "50" "54" "40" "32" ...
$ edu_master_graduated_f                     : chr [1:7760] "36" "36" "43" "48" ...
$ edu_master_ungraduated_m                   : chr [1:7760] "10" "10" "11" "6" ...
$ edu_master_ungraduated_f                   : chr [1:7760] "4" "12" "10" "7" ...
$ edu_university_graduated_m                 : chr [1:7760] "163" "149" "111" "129" ...
$ edu_university_graduated_f                 : chr [1:7760] "186" "204" "143" "147" ...
$ edu_university_ungraduated_m               : chr [1:7760] "55" "42" "32" "44" ...
$ edu_university_ungraduated_f               : chr [1:7760] "44" "30" "22" "19" ...
$ edu_juniorcollege_2ys_graduated_m          : chr [1:7760] "37" "37" "14" "28" ...
$ edu_juniorcollege_2ys_graduated_f          : chr [1:7760] "33" "43" "26" "43" ...
$ edu_juniorcollege_2ys_ungraduated_m        : chr [1:7760] "7" "3" "4" "3" ...
$ edu_juniorcollege_2ys_ungraduated_f        : chr [1:7760] "2" "5" "4" "4" ...
$ edu_juniorcollege_5ys_final2y_graduated_m  : chr [1:7760] "44" "46" "28" "30" ...
$ edu_juniorcollege_5ys_final2y_graduated_f  : chr [1:7760] "36" "45" "27" "25" ...
$ edu_juniorcollege_5ys_final2y_ungraduated_m: chr [1:7760] "0" "2" "1" "2" ...
$ edu_juniorcollege_5ys_final2y_ungraduated_f: chr [1:7760] "0" "2" "1" "0" ...
$ edu_senior_graduated_m                     : chr [1:7760] "71" "58" "30" "46" ...
$ edu_senior_graduated_f                     : chr [1:7760] "57" "49" "23" "52" ...
$ edu_senior_ungraduated_m                   : chr [1:7760] "19" "10" "7" "20" ...
$ edu_senior_ungraduated_f                   : chr [1:7760] "21" "14" "11" "22" ...
$ edu_seniorvocational_graduated_m           : chr [1:7760] "109" "72" "40" "71" ...
$ edu_seniorvocational_graduated_f           : chr [1:7760] "140" "100" "64" "82" ...
$ edu_seniorvocational_ungraduated_m         : chr [1:7760] "17" "14" "7" "4" ...
$ edu_seniorvocational_ungraduated_f         : chr [1:7760] "15" "8" "6" "8" ...
$ edu_juniorcollege_5ys_first3y_ungraduated_m: chr [1:7760] "4" "6" "2" "3" ...
$ edu_juniorcollege_5ys_first3y_ungraduated_f: chr [1:7760] "5" "6" "4" "2" ...
$ edu_junior_graduated_m                     : chr [1:7760] "42" "28" "27" "31" ...
$ edu_junior_graduated_f                     : chr [1:7760] "80" "43" "26" "35" ...
$ edu_junior_ungraduated_m                   : chr [1:7760] "5" "7" "4" "3" ...
$ edu_junior_ungraduated_f                   : chr [1:7760] "5" "11" "8" "6" ...
$ edu_juniorvocational_graduated_m           : chr [1:7760] "12" "19" "0" "1" ...
$ edu_juniorvocational_graduated_f           : chr [1:7760] "37" "24" "3" "1" ...
$ edu_juniorvocational_ungraduated_m         : chr [1:7760] "0" "0" "0" "0" ...
$ edu_juniorvocational_ungraduated_f         : chr [1:7760] "0" "0" "0" "0" ...
$ edu_primary_graduated_m                    : chr [1:7760] "26" "22" "14" "25" ...
$ edu_primary_graduated_f                    : chr [1:7760] "64" "58" "32" "52" ...
$ edu_primary_ungraduated_m                  : chr [1:7760] "2" "1" "0" "0" ...
$ edu_primary_ungraduated_f                  : chr [1:7760] "4" "3" "4" "4" ...
$ edu_selftaughtl_m                          : chr [1:7760] "0" "0" "0" "0" ...
$ edu_selftaughtl_f                          : chr [1:7760] "2" "0" "0" "1" ...
$ edu_illiterate_m                           : chr [1:7760] "0" "0" "1" "1" ...
$ edu_illiterate_f                           : chr [1:7760] "3" "3" "3" "2" ...
- attr(*, "spec")=
 .. cols(
 ..   statistic_yyy = col_character(),
 ..   district_code = col_character(),
 ..   site_id = col_character(),
 ..   village = col_character(),
 ..   edu_age_15up_total = col_character(),
 ..   edu_doctor_graduated_m = col_character(),
 ..   edu_doctor_graduated_f = col_character(),
 ..   edu_doctor_ungraduated_m = col_character(),
 ..   edu_doctor_ungraduated_f = col_character(),
 ..   edu_master_graduated_m = col_character(),
 ..   edu_master_graduated_f = col_character(),
 ..   edu_master_ungraduated_m = col_character(),
 ..   edu_master_ungraduated_f = col_character(),
 ..   edu_university_graduated_m = col_character(),
 ..   edu_university_graduated_f = col_character(),
 ..   edu_university_ungraduated_m = col_character(),
 ..   edu_university_ungraduated_f = col_character(),
 ..   edu_juniorcollege_2ys_graduated_m = col_character(),
 ..   edu_juniorcollege_2ys_graduated_f = col_character(),
 ..   edu_juniorcollege_2ys_ungraduated_m = col_character(),
 ..   edu_juniorcollege_2ys_ungraduated_f = col_character(),
 ..   edu_juniorcollege_5ys_final2y_graduated_m = col_character(),
 ..   edu_juniorcollege_5ys_final2y_graduated_f = col_character(),
 ..   edu_juniorcollege_5ys_final2y_ungraduated_m = col_character(),
 ..   edu_juniorcollege_5ys_final2y_ungraduated_f = col_character(),
 ..   edu_senior_graduated_m = col_character(),
 ..   edu_senior_graduated_f = col_character(),
 ..   edu_senior_ungraduated_m = col_character(),
 ..   edu_senior_ungraduated_f = col_character(),
 ..   edu_seniorvocational_graduated_m = col_character(),
 ..   edu_seniorvocational_graduated_f = col_character(),
 ..   edu_seniorvocational_ungraduated_m = col_character(),
 ..   edu_seniorvocational_ungraduated_f = col_character(),
 ..   edu_juniorcollege_5ys_first3y_ungraduated_m = col_character(),
 ..   edu_juniorcollege_5ys_first3y_ungraduated_f = col_character(),
 ..   edu_junior_graduated_m = col_character(),
 ..   edu_junior_graduated_f = col_character(),
 ..   edu_junior_ungraduated_m = col_character(),
 ..   edu_junior_ungraduated_f = col_character(),
 ..   edu_juniorvocational_graduated_m = col_character(),
 ..   edu_juniorvocational_graduated_f = col_character(),
 ..   edu_juniorvocational_ungraduated_m = col_character(),
 ..   edu_juniorvocational_ungraduated_f = col_character(),
 ..   edu_primary_graduated_m = col_character(),
 ..   edu_primary_graduated_f = col_character(),
 ..   edu_primary_ungraduated_m = col_character(),
 ..   edu_primary_ungraduated_f = col_character(),
 ..   edu_selftaughtl_m = col_character(),
 ..   edu_selftaughtl_f = col_character(),
 ..   edu_illiterate_m = col_character(),
 ..   edu_illiterate_f = col_character()
 .. )
}
\usage{
edu_village_original
}
\description{
edu_village_original
}
\keyword{datasets}
