% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_demo_distribution.R
\docType{data}
\name{income_demo_distribution}
\alias{income_demo_distribution}
\title{Data of Number of Income Recipients, Percentages of Income Recipients by Age and by Sex in Taiwan}
\format{
A tibble with 45 rows and 11 variables:
\describe{
  \item{year}{dbl year of data}
  \item{p_all}{dbl number of income recipients, 所得收入者人數合計}
  \item{p_under30}{dbl percentage of income recipients who are under 30, 未滿三十歲所得收入者人數比例}
  \item{p_30_34}{dbl percentage of income recipients who are 30-34, 三十至三十四歲所得收入者人數比例}
  \item{p_35_39}{dbl percentage of income recipients who are 35-39, 三十五至三十九歲所得收入者人數比例}
  \item{p_40_44}{dbl percentage of income recipients who are 40-44 , 四十至四十四歲所得收入者人數比例}
  \item{p_45_54}{dbl percentage of income recipients who are 45-54 , 四十五至五十四歲所得收入者人數比例}
  \item{p_55_64}{dbl percentage of income recipients who are 55-64 , 五十五至六十四歲所得收入者人數比例}
  \item{p_above65}{dbl percentage of income recipients who are above 65 , 六十五歲以上所得收入者人數比例}
  \item{p_m}{dbl percentage of income recipients who are male, 男性所得收入者人數比例}
  \item{p_f}{dbl percentage of income recipients who are female, 女性所得收入者人數比例}
}
}
\source{
\url{https://data.gov.tw/dataset/22230}

\url{https://win.dgbas.gov.tw/fies/a11.asp?year=106}
}
\usage{
income_demo_distribution
}
\description{
This data is in a wide format, from 1976 to 2020.
"家庭收支調查-所得收入者人數與按年齡組別及性別之分配"
}
\keyword{datasets}
