% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_income_town.R
\docType{data}
\name{individual_income_town}
\alias{individual_income_town}
\title{Data of individual income tax data by town in Taiwan}
\format{
A tibble with 2,576 rows and 11 variables:
\describe{
  \item{year}{dbl year of data}
  \item{county}{chr county(hsien/shih 縣/市 in Chinese) name}
  \item{town}{chr town(hsiang/chen/shih 鄉/鎮/市 in Chinese) name}
  \item{unit}{dbl taxing unit, 納稅單位}
  \item{individual_income}{dbl individual income, 綜合所得總額}
  \item{avg}{dbl mean of individual income, 綜合所得總額}
  \item{median}{dbl median of individual income}
  \item{q1}{dbl 1st quantile of individual income}
  \item{q3}{dbl 3rd quantile of individual income}
  \item{sd}{dbl standard deviation of individual income}
  \item{var}{dbl variance of individual income}
}
}
\source{
\url{https://data.gov.tw/dataset/103066}
}
\usage{
individual_income_town
}
\description{
This data is in a long format, from 101 to 107.
}
\keyword{datasets}
