% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/married_divorced_county.R
\docType{data}
\name{married_divorced_county}
\alias{married_divorced_county}
\title{Data of 2020 married related statistics by county in Taiwan}
\format{

}
\source{
\url{https://segis.moi.gov.tw/STAT/Web/Platform/QueryInterface/STAT_QueryProductView.aspx?pid=21E915C22091E0772D29E038496BF264&spid=7ED8D58E129BC680}

\url{https://segis.moi.gov.tw/STAT/Web/Platform/QueryInterface/STAT_QueryProductView.aspx?pid=FB03E79B61F018CA2AAED227D59E8523&spid=7ED8D58E129BC680}
}
\usage{
married_divorced_county
}
\description{
Data of 2020 married related statistics by county in Taiwan
}
\details{
These are combined household registration records collected from SEGIS
結合 SEGIS 的兩份資料，共包含:
(1) 結婚對數、結婚登記中外(含大陸港澳配偶)聯姻人數、結婚登記中外(含大陸港澳配偶)聯姻男性人數、結婚登記中外(含大陸港澳配偶)聯姻女性人數
(2) 離婚對數、離婚登記男性人數、離婚登記女性人數
(3) 備註： 此項中結婚部分統計不含『同婚』統計數值。
}
\keyword{datasets}
