% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscdata-package.R
\docType{package}
\name{pscdata-package}
\alias{pscdata}
\alias{pscdata-package}
\title{pscdata: Opendata for research - Population and Gender Studies}
\description{
Dataset for research in population and gender studies.
}
\keyword{internal}
