% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referendum_town.R
\docType{data}
\name{referendum_town}
\alias{referendum_town}
\title{Data of 2018 Taiwanese referendum - result of all proposals by town in Taiwan}
\format{
A tibble with 368 rows and 12 variables:
\describe{
  \item{countytown}{chr county(hsien/shih 縣/市 in Chinese) and town(hsiang/chen/shih 鄉/鎮/市 in Chinese) name}
  \item{countytown}{chr id of countytown}
  \item{ref7}{dbl denoting percentage of agree rate of referendum 07}
  \item{ref8}{dbl denoting percentage of agree rate of referendum 08}
  \item{ref9}{dbl denoting percentage of agree rate of referendum 09}
  \item{ref10}{dbl denoting percentage of agree rate of referendum 10}
  \item{ref11}{dbl denoting percentage of agree rate of referendum 11}
  \item{ref12}{dbl denoting percentage of agree rate of referendum 12}
  \item{ref13}{dbl denoting percentage of agree rate of referendum 13}
  \item{ref14}{dbl denoting percentage of agree rate of referendum 14}
  \item{ref15}{dbl denoting percentage of agree rate of referendum 15}
  \item{ref16}{dbl denoting percentage of agree rate of referendum 16}
}
}
\source{
\url{https://data.gov.tw/dataset/95883}
}
\usage{
referendum_town
}
\description{
These are combined town statistics and the result fo referendum in 2018
}
\keyword{datasets}
